Another whoops

+++ src/host/dmr/Slot.cpp
@@ -806,7 +806,7 @@ uint32_t Slot::getLastSrcId() const
     return 0U;
 }
 
-uint32_t Slot::getRssi() const
+uint8_t Slot::getRssi() const
 {
     return m_rssi;
 }
